По замечаниям. Превращает Main в функциональный компонент. Выносит константы.
